+++ feature/home/README.md
@@ -19,7 +19,12 @@ src/main/java/cx/aswin/boxlore/feature/home/
   HomeScreen.kt / HomeViewModel.kt / HomeViewModelAssembler.kt
   DiscoveryGreeting.kt / HomeListeningHistoryItem.kt
   DebugScreen.kt / DebugViewModel.kt / AdaptiveLearnerDebugSection.kt
-  components/ settings/ settings/pages/ settings/dialogs/ logic/
+  components/ settings/ settings/pages/ settings/dialogs/
+  logic/
+    PodcastAffinityLogic.kt
+    HomeDiscoveryLogic.kt      # adaptiveHistoryMaturityBucket, discoverPodcastsExcluding
+    HomeMappingLogic.kt      # Episode.toRecommendationPodcast
+    SerialEpisodeLogic.kt    # resolveNextSerialEpisode
 ```
 
 ## Dependencies
@@ -42,7 +47,7 @@ None owned. Uses DataStore / `BoxcastPrefs` / ranking DB through core modules â€
 ## Testing notes
 
 - Assemblers: `HomeViewModelAssembler`, `settings.SettingsViewModelAssembler`
-- JVM: `SettingsViewModelTest` (Turbine + fake RSS/ranking ports); `DiscoveryGreetingTest`; `logic/PodcastAffinityLogicTest`; `HomeListeningHistoryItemTest`
+- JVM: `SettingsViewModelTest` (Turbine + fake RSS/ranking ports); `DiscoveryGreetingTest`; `logic/PodcastAffinityLogicTest`, `logic/HomeDiscoveryLogicTest`, `logic/HomeMappingLogicTest`, `logic/SerialEpisodeLogicTest`; `HomeListeningHistoryItemTest`
 - Full Home VM construction still needs Application + heavy fakes (deferred)
 - **androidTest:** `AddRssFeedDialogUiTest`; `DownloadsSettingsPageUiTest`; composition smoke `AddRssFeedDialogScreenshotStubTest` (tags only; **not** a screenshot golden â€” P26 incomplete)
 - Stable `testTag`s: `home_settings_button`; `settings_add_rss_*`; `settings_downloads_smart` / `settings_downloads_auto`
