+++ feature/info/src/main/java/cx/aswin/boxlore/feature/info/PodcastInfoScreen.kt
@@ -141,6 +141,8 @@ import cx.aswin.boxlore.core.model.Episode
 import cx.aswin.boxlore.core.model.Person
 import cx.aswin.boxlore.core.model.Podcast
 import cx.aswin.boxlore.core.model.PodrollItem
+import cx.aswin.boxlore.feature.info.sections.PodcastInfoHeroSection
+import cx.aswin.boxlore.feature.info.sections.stripHtml
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.Flow
@@ -148,11 +150,6 @@ import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
-
-private fun stripHtml(html: String?): String {
-    if (html.isNullOrEmpty()) return ""
-    return Html.fromHtml(html, Html.FROM_HTML_MODE_LEGACY).toString().trim()
-}
 
 private fun extractDominantColor(bitmap: Bitmap): Color {
     val palette = Palette.from(bitmap).generate()
@@ -162,61 +159,6 @@ private fun extractDominantColor(bitmap: Bitmap): Color {
             ?: palette.dominantSwatch?.rgb
             ?: return Color.Transparent
     return Color(colorInt)
-}
-
-@Composable
-private fun CompactPersonChip(
-    person: Person,
-    onClick: () -> Unit,
-) {
-    Surface(
-        color = MaterialTheme.colorScheme.surfaceContainerHigh,
-        shape = ExpressiveShapes.Pill,
-        modifier =
-            Modifier.expressiveClickable(
-                enabled = !person.href.isNullOrBlank(),
-                onClick = onClick,
-            ),
-    ) {
-        Row(
-            modifier = Modifier.padding(horizontal = 12.dp, vertical = 8.dp),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.spacedBy(4.dp),
-        ) {
-            if (!person.img.isNullOrBlank()) {
-                OptimizedImage(
-                    url = person.img,
-                    proxyWidth = 40,
-                    contentDescription = person.name,
-                    modifier =
-                        Modifier
-                            .size(16.dp)
-                            .clip(CircleShape),
-                    contentScale = ContentScale.Crop,
-                )
-            } else {
-                Icon(
-                    imageVector = Icons.Rounded.Person,
-                    contentDescription = null,
-                    modifier = Modifier.size(16.dp),
-                    tint = MaterialTheme.colorScheme.onSurface,
-                )
-            }
-
-            val displayText =
-                if (!person.role.isNullOrBlank()) {
-                    "${person.name} (${person.role!!.replaceFirstChar { it.uppercaseChar() }})"
-                } else {
-                    person.name
-                }
-            Text(
-                text = displayText,
-                style = MaterialTheme.typography.labelMedium,
-                fontWeight = FontWeight.Bold,
-                color = MaterialTheme.colorScheme.onSurface,
-            )
-        }
-    }
 }
 
 // Navbar height constant
@@ -1006,780 +948,8 @@ private fun resolveAutoScrollTarget(
 
 // endregion
 
-// region PodcastInfoScreen extraction: update frequency / genre / medium chip helpers
 
-private fun calculateUpdateFrequencyData(
-    podcast: Podcast,
-    episodes: List<Episode>,
-    currentSort: EpisodeSort,
-    searchQuery: String,
-    cachedFrequencyData: Pair<String, ImageVector>?,
-): Pair<String, ImageVector>? {
-    // If sorted by oldest or searching, use the cached calculation (to prevent "Inactive / Ended" bug)
-    if (currentSort != EpisodeSort.NEWEST || searchQuery.isNotEmpty()) {
-        return cachedFrequencyData
-    }
 
-    // 1. Securely sort and take the latest 15 episodes (Recent History)
-    val validEpisodes = filterValidFrequencyEpisodes(episodes)
-    val daysSinceLatest = computeDaysSinceLatest(podcast, validEpisodes)
-
-    // 2. Check if it's dead or on hiatus
-    dormancyStatus(daysSinceLatest)?.let { return it }
-
-    // 3. Check for decay / delayed seasons (Between Seasons check)
-    val medianIntervalDays = computeMedianIntervalDays(validEpisodes, podcast)
-    betweenSeasonsStatus(medianIntervalDays, daysSinceLatest)?.let { return it }
-
-    // 4. Use the explicit updateFrequency tag if available
-    explicitFrequencyTag(podcast.updateFrequency)?.let { return it }
-
-    // 5. Fallback: Predict frequency using Median and Day of Week counts
-    if (validEpisodes.size < 4 || medianIntervalDays == null) return cachedFrequencyData
-    return predictedFrequencyFromPattern(validEpisodes, medianIntervalDays) ?: cachedFrequencyData
-}
-
-private fun filterValidFrequencyEpisodes(episodes: List<Episode>): List<Episode> =
-    episodes
-        .filter { it.episodeType != "trailer" && it.episodeType != "bonus" && it.publishedDate > 0 }
-        .sortedByDescending { it.publishedDate }
-        .take(15)
-
-private fun computeDaysSinceLatest(
-    podcast: Podcast,
-    validEpisodes: List<Episode>,
-): Long? {
-    val latestEpisodeDate = validEpisodes.firstOrNull()?.publishedDate ?: podcast.latestEpisode?.publishedDate
-    return latestEpisodeDate?.let { (System.currentTimeMillis() / 1000 - it) / (60 * 60 * 24) }
-}
-
-private fun dormancyStatus(daysSinceLatest: Long?): Pair<String, ImageVector>? {
-    if (daysSinceLatest == null || daysSinceLatest <= 0) return null
-    return when {
-        daysSinceLatest > 365 -> Pair("Inactive / Ended", Icons.Rounded.PauseCircle)
-        daysSinceLatest > 180 -> Pair("On Hiatus", Icons.Rounded.PauseCircle)
-        else -> null
-    }
-}
-
-/** Median gap (days) between the latest episodes, falling back to the feed's declared tag. */
-private fun computeMedianIntervalDays(
-    validEpisodes: List<Episode>,
-    podcast: Podcast,
-): Long? {
-    if (validEpisodes.size >= 4) {
-        val intervals = mutableListOf<Long>()
-        for (i in 0 until validEpisodes.size - 1) {
-            val newer = validEpisodes[i].publishedDate
-            val older = validEpisodes[i + 1].publishedDate
-            val daysDiff = (newer - older) / (60 * 60 * 24)
-            if (daysDiff >= 0) intervals.add(daysDiff)
-        }
-        if (intervals.isEmpty()) return null
-        val sortedIntervals = intervals.sorted()
-        return sortedIntervals[sortedIntervals.size / 2]
-    }
-    // Estimate based on tag if episodes are scarce
-    val tag = podcast.updateFrequency?.lowercase() ?: ""
-    return when {
-        tag.contains("daily") -> 1L
-        tag.contains("weekly") -> 7L
-        tag.contains("bi-weekly") || tag.contains("2 weeks") -> 14L
-        tag.contains("monthly") -> 30L
-        else -> null
-    }
-}
-
-private fun betweenSeasonsStatus(
-    medianIntervalDays: Long?,
-    daysSinceLatest: Long?,
-): Pair<String, ImageVector>? {
-    if (medianIntervalDays == null || medianIntervalDays <= 3 || daysSinceLatest == null) return null
-    return if (daysSinceLatest > (medianIntervalDays * 2)) {
-        Pair("Between Seasons", Icons.Rounded.HourglassBottom)
-    } else {
-        null
-    }
-}
-
-private fun explicitFrequencyTag(tag: String?): Pair<String, ImageVector>? {
-    if (tag.isNullOrBlank()) return null
-    val cleanText = tag.trim().lowercase()
-    val parsedDouble = cleanText.toDoubleOrNull()
-    val formattedText =
-        if (parsedDouble != null) {
-            when {
-                parsedDouble >= 7.0 -> "Releases Daily"
-                parsedDouble >= 2.0 -> "Releases Multi-Weekly"
-                parsedDouble >= 1.0 -> "Releases Weekly"
-                parsedDouble >= 0.5 -> "Releases Every 2 Weeks"
-                parsedDouble >= 0.1 -> "Releases Monthly"
-                else -> "Releases Occasionally"
-            }
-        } else {
-            when (cleanText) {
-                "daily" -> "Releases Daily"
-                "weekly" -> "Releases Weekly"
-                "monthly" -> "Releases Monthly"
-                "biweekly", "bi-weekly" -> "Releases Every 2 Weeks"
-                else -> tag.replaceFirstChar { if (it.isLowerCase()) it.titlecase() else it.toString() }
-            }
-        }
-    val icon = if (formattedText.contains("Daily", ignoreCase = true)) Icons.Rounded.Bolt else Icons.Rounded.CalendarMonth
-    return Pair(formattedText, icon)
-}
-
-/** Most common weekday name across [validEpisodes], if at least half of them share one. */
-private fun commonReleaseDayName(validEpisodes: List<Episode>): String? {
-    val calendar = java.util.Calendar.getInstance()
-    val dayCounts = IntArray(8)
-    for (ep in validEpisodes) {
-        calendar.timeInMillis = ep.publishedDate * 1000
-        dayCounts[calendar.get(java.util.Calendar.DAY_OF_WEEK)]++
-    }
-    var maxDay = -1
-    var maxCount = 0
-    for (i in 1..7) {
-        if (dayCounts[i] > maxCount) {
-            maxCount = dayCounts[i]
-            maxDay = i
-        }
-    }
-    if (maxCount < (validEpisodes.size * 0.5).toInt()) return null
-    return when (maxDay) {
-        java.util.Calendar.SUNDAY -> "Sundays"
-        java.util.Calendar.MONDAY -> "Mondays"
-        java.util.Calendar.TUESDAY -> "Tuesdays"
-        java.util.Calendar.WEDNESDAY -> "Wednesdays"
-        java.util.Calendar.THURSDAY -> "Thursdays"
-        java.util.Calendar.FRIDAY -> "Fridays"
-        java.util.Calendar.SATURDAY -> "Saturdays"
-        else -> null
-    }
-}
-
-private fun predictedFrequencyFromPattern(
-    validEpisodes: List<Episode>,
-    medianIntervalDays: Long,
-): Pair<String, ImageVector>? {
-    val commonDayName = commonReleaseDayName(validEpisodes)
-    return when (medianIntervalDays) {
-        in 0..1 -> Pair("Releases Daily", Icons.Rounded.Bolt)
-        in 2..4 -> Pair("Releases Multi-Weekly", Icons.Rounded.CalendarMonth)
-        in 5..8 ->
-            Pair(
-                if (commonDayName != null) "Weekly on $commonDayName" else "Releases Weekly",
-                Icons.Rounded.CalendarMonth,
-            )
-        in 12..16 ->
-            Pair(
-                if (commonDayName != null) "Every 2 Weeks on $commonDayName" else "Releases Every 2 Weeks",
-                Icons.Rounded.CalendarMonth,
-            )
-        in 25..35 -> Pair("Releases Monthly", Icons.Rounded.CalendarMonth)
-        else -> null
-    }
-}
-
-private fun genreIconFor(genre: String): ImageVector {
-    val genreLc = genre.lowercase()
-    return when {
-        genreLc.contains("music") -> Icons.Rounded.MusicNote
-        genreLc.contains("comedy") -> Icons.Rounded.SentimentVerySatisfied
-        genreLc.contains("sport") -> Icons.Rounded.EmojiEvents
-        genreLc.contains("science") -> Icons.Rounded.Science
-        genreLc.contains("tech") -> Icons.Rounded.Computer
-        genreLc.contains("news") -> Icons.Rounded.Newspaper
-        genreLc.contains("health") -> Icons.Rounded.MonitorHeart
-        genreLc.contains("history") -> Icons.Rounded.AccountBalance
-        genreLc.contains("arts") -> Icons.Rounded.Palette
-        genreLc.contains("education") -> Icons.Rounded.School
-        genreLc.contains("tv") || genreLc.contains("film") -> Icons.Rounded.Movie
-        genreLc.contains("fiction") -> Icons.Rounded.AutoStories
-        genreLc.contains("religion") || genreLc.contains("spiritual") -> Icons.Rounded.SelfImprovement
-        genreLc.contains("family") || genreLc.contains("kids") -> Icons.Rounded.ChildCare
-        genreLc.contains("leisure") -> Icons.Rounded.Weekend
-        genreLc.contains("business") -> Icons.Rounded.Work
-        genreLc.contains("government") -> Icons.Rounded.Gavel
-        genreLc.contains("society") || genreLc.contains("culture") -> Icons.Rounded.Groups
-        genreLc.contains("crime") -> Icons.Rounded.Fingerprint
-        else -> Icons.Rounded.Category
-    }
-}
-
-private fun mediumIconFor(medium: String): ImageVector =
-    when (medium.lowercase()) {
-        "music" -> Icons.Rounded.MusicNote
-        "video" -> Icons.Rounded.Videocam
-        "film" -> Icons.Rounded.Movie
-        "audiobook" -> Icons.Rounded.AutoStories
-        "newsletter" -> Icons.Rounded.Email
-        "blog" -> Icons.AutoMirrored.Rounded.Article
-        else -> Icons.Rounded.Headphones
-    }
-
-// endregion
-
-// region PodcastInfoScreen extraction: metadata chips row
-
-@Composable
-private fun RssFeedChip() {
-    Surface(
-        shape = ExpressiveShapes.Pill,
-        color = MaterialTheme.colorScheme.surfaceContainerHigh,
-    ) {
-        Row(
-            modifier = Modifier.padding(horizontal = 12.dp, vertical = 8.dp),
-            horizontalArrangement = Arrangement.spacedBy(6.dp),
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            Icon(
-                imageVector = Icons.Rounded.RssFeed,
-                contentDescription = null,
-                modifier = Modifier.size(16.dp),
-                tint = MaterialTheme.colorScheme.primary,
-            )
-            Text(
-                text = "RSS",
-                style = MaterialTheme.typography.labelMedium,
-                fontWeight = FontWeight.Medium,
-                color = MaterialTheme.colorScheme.primary,
-            )
-        }
-    }
-}
-
-@Composable
-private fun UpdateFrequencyChip(frequencyData: Pair<String, ImageVector>) {
-    Surface(
-        shape = ExpressiveShapes.Pill,
-        color = MaterialTheme.colorScheme.surfaceContainerHigh,
-    ) {
-        Row(
-            modifier = Modifier.padding(horizontal = 12.dp, vertical = 8.dp),
-            horizontalArrangement = Arrangement.spacedBy(6.dp),
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            Icon(
-                imageVector = frequencyData.second,
-                contentDescription = null,
-                modifier = Modifier.size(16.dp),
-                tint = MaterialTheme.colorScheme.primary,
-            )
-            Text(
-                text = frequencyData.first,
-                style = MaterialTheme.typography.labelMedium,
-                fontWeight = FontWeight.Medium,
-                color = MaterialTheme.colorScheme.primary,
-            )
-        }
-    }
-}
-
-@Composable
-private fun GenreChip(genre: String) {
-    Surface(
-        shape = ExpressiveShapes.Pill,
-        color = MaterialTheme.colorScheme.secondaryContainer,
-    ) {
-        Row(
-            modifier = Modifier.padding(horizontal = 12.dp, vertical = 8.dp),
-            horizontalArrangement = Arrangement.spacedBy(4.dp),
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            Icon(
-                imageVector = genreIconFor(genre),
-                contentDescription = null,
-                modifier = Modifier.size(16.dp),
-                tint = MaterialTheme.colorScheme.onSecondaryContainer,
-            )
-            Text(
-                text = genre,
-                style = MaterialTheme.typography.labelMedium,
-                color = MaterialTheme.colorScheme.onSecondaryContainer,
-                fontWeight = FontWeight.Bold,
-            )
-        }
-    }
-}
-
-@Composable
-private fun PlayTrailerChip(onClick: () -> Unit) {
-    Surface(
-        shape = ExpressiveShapes.Pill,
-        color = MaterialTheme.colorScheme.primary,
-        modifier = Modifier.expressiveClickable(onClick = onClick),
-    ) {
-        Row(
-            modifier = Modifier.padding(horizontal = 12.dp, vertical = 8.dp),
-            horizontalArrangement = Arrangement.spacedBy(4.dp),
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            Icon(
-                imageVector = Icons.Rounded.PlayArrow,
-                contentDescription = null,
-                modifier = Modifier.size(16.dp),
-                tint = MaterialTheme.colorScheme.onPrimary,
-            )
-            Text(
-                text = "Play Trailer",
-                style = MaterialTheme.typography.labelMedium,
-                color = MaterialTheme.colorScheme.onPrimary,
-                fontWeight = FontWeight.Bold,
-            )
-        }
-    }
-}
-
-@Composable
-private fun MediumChip(medium: String) {
-    Surface(
-        shape = ExpressiveShapes.Pill,
-        color = MaterialTheme.colorScheme.primaryContainer,
-    ) {
-        Row(
-            modifier = Modifier.padding(horizontal = 12.dp, vertical = 8.dp),
-            horizontalArrangement = Arrangement.spacedBy(4.dp),
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            Icon(
-                imageVector = mediumIconFor(medium),
-                contentDescription = null,
-                modifier = Modifier.size(16.dp),
-                tint = MaterialTheme.colorScheme.onPrimaryContainer,
-            )
-            Text(
-                text = medium.replaceFirstChar { c -> c.uppercaseChar() },
-                style = MaterialTheme.typography.labelMedium,
-                color = MaterialTheme.colorScheme.onPrimaryContainer,
-                fontWeight = FontWeight.Bold,
-            )
-        }
-    }
-}
-
-@Composable
-private fun FundingChip(
-    fundingMessage: String?,
-    onClick: () -> Unit,
-) {
-    Surface(
-        shape = ExpressiveShapes.Pill,
-        color = MaterialTheme.colorScheme.tertiaryContainer,
-        modifier = Modifier.expressiveClickable(onClick = onClick),
-    ) {
-        Row(
-            modifier = Modifier.padding(horizontal = 12.dp, vertical = 8.dp),
-            horizontalArrangement = Arrangement.spacedBy(4.dp),
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            Icon(
-                imageVector = Icons.Filled.Favorite,
-                contentDescription = null,
-                modifier = Modifier.size(16.dp),
-                tint = MaterialTheme.colorScheme.onTertiaryContainer,
-            )
-            Text(
-                text = fundingMessage ?: "Support",
-                style = MaterialTheme.typography.labelMedium,
-                color = MaterialTheme.colorScheme.onTertiaryContainer,
-                fontWeight = FontWeight.Bold,
-            )
-        }
-    }
-}
-
-@Composable
-private fun PodcastInfoMetadataChipsRow(
-    podcast: Podcast,
-    sortedPersons: List<Person>,
-    trailerEpisode: Episode?,
-    frequencyData: Pair<String, ImageVector>?,
-    context: android.content.Context,
-    onPlayTrailer: (Episode) -> Unit,
-) {
-    val medium = resolveDisplayMedium(podcast)
-
-    LazyRow(
-        horizontalArrangement = Arrangement.spacedBy(8.dp, Alignment.CenterHorizontally),
-        contentPadding = PaddingValues(horizontal = 0.dp),
-        modifier = Modifier.fillMaxWidth(),
-    ) {
-        podcastIdentityChips(podcast, frequencyData)
-
-        items(sortedPersons) { person ->
-            CompactPersonChip(person = person, onClick = { openPersonLink(context, person) })
-        }
-
-        trailerAndMediumChips(trailerEpisode, medium, onPlayTrailer)
-        fundingChip(podcast, context)
-    }
-}
-
-/** "video" if this is a video-only podcast feed masquerading as "podcast" medium; otherwise unchanged. */
-private fun resolveDisplayMedium(podcast: Podcast): String? =
-    if (podcast.medium == "podcast" && podcast.latestEpisode?.enclosureType?.startsWith("video/") == true) {
-        "video"
-    } else {
-        podcast.medium
-    }
-
-private fun openPersonLink(
-    context: android.content.Context,
-    person: Person,
-) {
-    if (person.href.isNullOrBlank()) return
-    context.startActivity(Intent(Intent.ACTION_VIEW, Uri.parse(person.href)))
-}
-
-private fun LazyListScope.podcastIdentityChips(
-    podcast: Podcast,
-    frequencyData: Pair<String, ImageVector>?,
-) {
-    if (podcast.isRss) {
-        item { RssFeedChip() }
-    }
-    if (frequencyData != null) {
-        item { UpdateFrequencyChip(frequencyData) }
-    }
-    if (podcast.genre.isNotEmpty()) {
-        item { GenreChip(podcast.genre) }
-    }
-}
-
-private fun LazyListScope.trailerAndMediumChips(
-    trailerEpisode: Episode?,
-    medium: String?,
-    onPlayTrailer: (Episode) -> Unit,
-) {
-    if (trailerEpisode != null) {
-        item { PlayTrailerChip(onClick = { onPlayTrailer(trailerEpisode) }) }
-    }
-    if (!medium.isNullOrEmpty() && medium != "podcast") {
-        item { MediumChip(medium) }
-    }
-}
-
-private fun LazyListScope.fundingChip(
-    podcast: Podcast,
-    context: android.content.Context,
-) {
-    if (podcast.fundingUrl == null) return
-    item {
-        FundingChip(
-            fundingMessage = podcast.fundingMessage,
-            onClick = {
-                val intent = Intent(Intent.ACTION_VIEW, Uri.parse(podcast.fundingUrl))
-                context.startActivity(intent)
-            },
-        )
-    }
-}
-
-// endregion
-
-// region PodcastInfoScreen extraction: description section
-
-@Composable
-private fun LockedFeedNotice() {
-    var showLockedInfoDialog by remember { mutableStateOf(false) }
-
-    Spacer(modifier = Modifier.height(12.dp))
-    androidx.compose.material3.HorizontalDivider(
-        thickness = 0.5.dp,
-        color = MaterialTheme.colorScheme.outlineVariant.copy(alpha = 0.4f),
-    )
-    Spacer(modifier = Modifier.height(8.dp))
-
-    Row(
-        modifier =
-            Modifier
-                .fillMaxWidth()
-                .clip(MaterialTheme.shapes.small)
-                .expressiveClickable(isolate = true) { showLockedInfoDialog = true }
-                .padding(vertical = 4.dp),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.spacedBy(6.dp),
-    ) {
-        Icon(
-            imageVector = Icons.Rounded.Lock,
-            contentDescription = "Locked",
-            tint = MaterialTheme.colorScheme.error,
-            modifier = Modifier.size(16.dp),
-        )
-        Text(
-            text = "Podcast feed is locked",
-            style = MaterialTheme.typography.labelMedium,
-            fontWeight = FontWeight.Bold,
-            color = MaterialTheme.colorScheme.error,
-        )
-        Icon(
-            imageVector = Icons.Rounded.Info,
-            contentDescription = "Information",
-            tint = MaterialTheme.colorScheme.error.copy(alpha = 0.7f),
-            modifier = Modifier.size(14.dp),
-        )
-    }
-
-    if (showLockedInfoDialog) {
-        AlertDialog(
-            onDismissRequest = { showLockedInfoDialog = false },
-            icon = {
-                Icon(
-                    imageVector = Icons.Rounded.Lock,
-                    contentDescription = null,
-                    tint = MaterialTheme.colorScheme.error,
-                )
-            },
-            title = {
-                Text(
-                    text = "Podcast Locked",
-                    style = MaterialTheme.typography.titleMedium,
-                    fontWeight = FontWeight.Bold,
-                )
-            },
-            text = {
-                Text(
-                    text = "This podcast's feed has been locked by its publisher. According to the Podcasting 2.0 specification, a locked feed prevents other directory platforms or hosting services from importing or migrating this show's feed without the owner's explicit authorization.",
-                    style = MaterialTheme.typography.bodyMedium,
-                )
-            },
-            confirmButton = {
-                TextButton(onClick = { showLockedInfoDialog = false }) {
-                    Text("Got it")
-                }
-            },
-        )
-    }
-}
-
-@Composable
-private fun PodrollRecommendations(
-    podroll: List<PodrollItem>,
-    onPodcastClick: (String) -> Unit,
-) {
-    Spacer(modifier = Modifier.height(12.dp))
-    androidx.compose.material3.HorizontalDivider(
-        thickness = 0.5.dp,
-        color = MaterialTheme.colorScheme.outlineVariant.copy(alpha = 0.4f),
-    )
-    Spacer(modifier = Modifier.height(12.dp))
-
-    Text(
-        text = "Creator Recommends",
-        style = MaterialTheme.typography.labelLarge,
-        fontWeight = FontWeight.Bold,
-        color = MaterialTheme.colorScheme.onSurfaceVariant,
-    )
-
-    Spacer(modifier = Modifier.height(12.dp))
-
-    LazyRow(
-        modifier = Modifier.fillMaxWidth(),
-        horizontalArrangement = Arrangement.spacedBy(8.dp),
-    ) {
-        items(podroll) { item ->
-            RecommendedPodcastCard(
-                item = item,
-                onPodcastClick = onPodcastClick,
-            )
-        }
-    }
-}
-
-@Composable
-private fun PodcastInfoDescriptionSection(
-    strippedDesc: String,
-    isLocked: Boolean,
-    podroll: List<PodrollItem>?,
-    isDescExpanded: Boolean,
-    onToggleExpanded: () -> Unit,
-    onPodcastClick: (String) -> Unit,
-) {
-    val hasPodroll = !podroll.isNullOrEmpty()
-    if (strippedDesc.isEmpty() && !isLocked && !hasPodroll) return
-
-    Surface(
-        modifier =
-            Modifier
-                .fillMaxWidth()
-                .animateContentSize(
-                    animationSpec =
-                        spring(
-                            dampingRatio = Spring.DampingRatioLowBouncy,
-                            stiffness = Spring.StiffnessMediumLow,
-                        ),
-                ).expressiveClickable { onToggleExpanded() },
-        color = MaterialTheme.colorScheme.surfaceContainerLow,
-        shape = MaterialTheme.shapes.large,
-    ) {
-        PodcastInfoDescriptionBody(
-            strippedDesc = strippedDesc,
-            isLocked = isLocked,
-            podroll = podroll,
-            hasPodroll = hasPodroll,
-            isDescExpanded = isDescExpanded,
-            onPodcastClick = onPodcastClick,
-        )
-    }
-}
-
-@Composable
-private fun PodcastInfoDescriptionBody(
-    strippedDesc: String,
-    isLocked: Boolean,
-    podroll: List<PodrollItem>?,
-    hasPodroll: Boolean,
-    isDescExpanded: Boolean,
-    onPodcastClick: (String) -> Unit,
-) {
-    Column(
-        modifier =
-            Modifier
-                .fillMaxWidth()
-                .padding(16.dp),
-    ) {
-        if (strippedDesc.isNotEmpty()) {
-            Text(
-                text = strippedDesc,
-                style = MaterialTheme.typography.bodyMedium,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-                maxLines = if (isDescExpanded) Int.MAX_VALUE else 2,
-                overflow = TextOverflow.Ellipsis,
-                lineHeight = 20.sp,
-                modifier = Modifier.fillMaxWidth(),
-            )
-        }
-
-        // With no description there's nothing to "expand" — always show the lock notice and
-        // podroll in that case instead of hiding them behind a collapsed, empty description.
-        val showMetadataRegardless = isDescExpanded || strippedDesc.isEmpty()
-        if (showMetadataRegardless && isLocked) {
-            LockedFeedNotice()
-        }
-        if (showMetadataRegardless && hasPodroll) {
-            PodrollRecommendations(podroll = podroll.orEmpty(), onPodcastClick = onPodcastClick)
-        }
-    }
-}
-
-// endregion
-
-// region PodcastInfoScreen extraction: hero section
-
-@Composable
-private fun PodcastInfoHeroSection(
-    state: PodcastInfoUiState.Success,
-    sortedPersons: List<Person>,
-    isDescExpanded: Boolean,
-    onDescExpandedChange: (Boolean) -> Unit,
-    onPlayEpisode: (Episode) -> Unit,
-    onPodcastClick: (String) -> Unit,
-    context: android.content.Context,
-) {
-    Column(
-        modifier =
-            Modifier
-                .fillMaxWidth()
-                .padding(horizontal = 16.dp),
-        horizontalAlignment = Alignment.CenterHorizontally,
-    ) {
-        // 1. Centered Large Image
-        Surface(
-            modifier = Modifier.size(180.dp),
-            shape = MaterialTheme.shapes.extraLarge,
-            shadowElevation = 8.dp,
-        ) {
-            OptimizedImage(
-                url = state.podcast.imageUrl.takeIf { it.isNotEmpty() } ?: state.podcast.fallbackImageUrl,
-                proxyWidth = 600, // 180dp * ~3x density
-                contentDescription = state.podcast.title,
-                modifier = Modifier.fillMaxSize(),
-                contentScale = ContentScale.Crop,
-            )
-        }
-
-        Spacer(modifier = Modifier.height(20.dp))
-
-        // 2. Title & Artist
-        Text(
-            text = state.podcast.title,
-            style = MaterialTheme.typography.headlineMedium,
-            color = MaterialTheme.colorScheme.onBackground,
-            fontWeight = FontWeight.Bold,
-            textAlign = TextAlign.Center,
-            modifier = Modifier.padding(horizontal = 16.dp),
-        )
-
-        Spacer(modifier = Modifier.height(4.dp))
-
-        Text(
-            text = state.podcast.artist,
-            style = MaterialTheme.typography.titleMedium,
-            color = MaterialTheme.colorScheme.onSurfaceVariant,
-            fontWeight = FontWeight.Medium,
-            textAlign = TextAlign.Center,
-            modifier = Modifier.padding(horizontal = 16.dp),
-        )
-
-        Spacer(modifier = Modifier.height(12.dp))
-
-        // Update Frequency Calculation — filters/sorts episodes and analyzes release intervals,
-        // which can be expensive for large RSS catalogs, so it runs off the composition thread
-        // and is cancelled/restarted automatically as its keys change.
-        var cachedFrequencyData by remember(state.podcast.id) { mutableStateOf<Pair<String, ImageVector>?>(null) }
-
-        val frequencyData by produceState(
-            initialValue = cachedFrequencyData,
-            state.podcast,
-            state.episodes,
-            state.currentSort,
-            state.searchQuery,
-        ) {
-            val computed =
-                withContext(Dispatchers.Default) {
-                    calculateUpdateFrequencyData(
-                        podcast = state.podcast,
-                        episodes = state.episodes,
-                        currentSort = state.currentSort,
-                        searchQuery = state.searchQuery,
-                        cachedFrequencyData = cachedFrequencyData,
-                    )
-                }
-            cachedFrequencyData = computed
-            value = computed
-        }
-
-        // 3. Scrollable Metadata Chips Row — centered
-        val trailerEpisode =
-            remember(state.episodes) {
-                state.episodes.firstOrNull { it.episodeType == "trailer" }
-            }
-
-        PodcastInfoMetadataChipsRow(
-            podcast = state.podcast,
-            sortedPersons = sortedPersons,
-            trailerEpisode = trailerEpisode,
-            frequencyData = frequencyData,
-            context = context,
-            onPlayTrailer = onPlayEpisode,
-        )
-
-        Spacer(modifier = Modifier.height(18.dp))
-
-        val strippedDesc = remember(state.podcast.description) { stripHtml(state.podcast.description) }
-        PodcastInfoDescriptionSection(
-            strippedDesc = strippedDesc,
-            isLocked = state.podcast.isLocked,
-            podroll = state.podcast.podroll,
-            isDescExpanded = isDescExpanded,
-            onToggleExpanded = { onDescExpandedChange(!isDescExpanded) },
-            onPodcastClick = onPodcastClick,
-        )
-    }
-}
-
-// endregion
 
 // region PodcastInfoScreen extraction: toolbar notification/auto-download handlers
 
@@ -3533,51 +2703,6 @@ fun TrailerStackCard(
     }
 }
 
-@Composable
-private fun RecommendedPodcastCard(
-    item: cx.aswin.boxlore.core.model.PodrollItem,
-    onPodcastClick: (String) -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    Surface(
-        modifier =
-            modifier
-                .expressiveClickable(isolate = true) {
-                    val targetId =
-                        if (!item.uuid.isNullOrBlank()) {
-                            "guid:${item.uuid}"
-                        } else {
-                            val encoded = Uri.encode(item.url)
-                            "url:$encoded"
-                        }
-                    onPodcastClick(targetId)
-                },
-        shape = MaterialTheme.shapes.medium,
-        color = MaterialTheme.colorScheme.surfaceContainerHigh,
-        border = BorderStroke(0.5.dp, MaterialTheme.colorScheme.outlineVariant),
-    ) {
-        Row(
-            modifier = Modifier.padding(horizontal = 12.dp, vertical = 10.dp),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.spacedBy(8.dp),
-        ) {
-            Icon(
-                imageVector = Icons.Rounded.Podcasts,
-                contentDescription = null,
-                modifier = Modifier.size(18.dp),
-                tint = MaterialTheme.colorScheme.onSurfaceVariant,
-            )
-            Text(
-                text = item.title,
-                style = MaterialTheme.typography.bodyMedium,
-                fontWeight = FontWeight.Medium,
-                color = MaterialTheme.colorScheme.onSurface,
-                maxLines = 1,
-                overflow = TextOverflow.Ellipsis,
-            )
-        }
-    }
-}
 
 @Composable
 fun EpisodePlayStateWrapper(
